pyproject.toml
[project]
name = "zpacecode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.5.8",
    "anthropic>=0.84.0",
    "fastapi>=0.135.1",
    "html-to-markdown>=2.28.0",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "lxml-stubs>=0.5.1",
]
